Create LanguageSelector component; update Cypress tests.
