+++ cypress/e2e/participant/playSWR.cy.js
@@ -5,13 +5,7 @@ describe('Testing playthrough of SWR as a participant', () => {
     cy.login(Cypress.env('participantUsername'), Cypress.env('participantPassword'));
     cy.visit('/');
 
-    cy.get('.p-dropdown-trigger', { timeout: 10 * timeout })
-      .should('be.visible')
-      .click();
-    cy.get('.p-dropdown-item', { timeout: 10 * timeout })
-      .contains(Cypress.env('testRoarAppsAdministration'))
-      .should('be.visible')
-      .click();
+    cy.selectAdministration(Cypress.env('testRoarAppsAdministration'));
 
     cy.get('.p-tabview').contains('ROAR-Word');
     cy.visit(`/game/swr`);
