+++ cypress/support/commands.js
@@ -75,6 +75,16 @@ Cypress.Commands.add('activateAdminSidebar', () => {
   cy.get('[data-cy="button-admin-sidebar"]').click().wait(1000);
 });
 
+Cypress.Commands.add('selectAdministration', (testAdministration) => {
+  cy.get('[data-cy="dropdown-select-administration"]', { timeout: 2 * Cypress.env('timeout') })
+    .should('be.visible')
+    .click();
+  cy.get('.p-dropdown-item', { timeout: 2 * Cypress.env('timeout') })
+    .contains(testAdministration)
+    .should('be.visible')
+    .click();
+});
+
 Cypress.Commands.add('getAdministrationCard', (testAdministration, sort = 'descending') => {
   cy.get('[data-cy="dropdown-sort-administrations"]', { timeout: Cypress.env('timeout') }).click();
   cy.get('ul > li').contains(`Name (${sort})`).click();
