+++ src/components/NavBar.vue
@@ -8,32 +8,7 @@
       </router-link>
 
       <div id="navBarRightEnd" class="flex flex-row align-items-center">
-        <PvDropdown
-          v-model="$i18n.locale"
-          class="mr-2"
-          :options="languageDropdownOptions"
-          option-label="name"
-          option-value="value"
-          placeholder="Select language"
-          :highlight-on-select="true"
-        >
-          <template #value="locale">
-            <div v-if="locale.value" class="flex flex-row justify-content-center align-items-center">
-              <country-flag :country="getCountryFlag(locale.value)" class="mr-2" size="small" />
-              <span>{{ getCountryName(locale.value) }}</span>
-            </div>
-            <span v-else>
-              {{ locale.placeholder }}
-            </span>
-          </template>
-          <template #option="country">
-            <div class="flex flex-row justify-content-start align-items-center">
-              <country-flag :country="country.option.code" class="mr-2" size="small" />
-              <span>{{ country.option.name }}</span>
-            </div>
-          </template>
-        </PvDropdown>
-
+        <LanguageSelector />
         <div class="login-container">
           <div v-if="isAdmin">
             <PvButton label="Menu" icon="pi pi-bars" @click="toggleMenu" />
@@ -65,8 +40,6 @@ import _union from 'lodash/union';
 import { getSidebarActions } from '@/router/sidebarActions';
 import { fetchDocById } from '@/helpers/query/utils';
 import { useQuery } from '@tanstack/vue-query';
-import { languageOptions } from '../translations/i18n.js';
-import CountryFlag from 'vue-country-flag-next';
 
 const router = useRouter();
 const authStore = useAuthStore();
@@ -107,33 +80,6 @@ const isSuperAdmin = computed(() => Boolean(userClaims.value?.claims?.super_admi
 const isAtHome = computed(() => {
   return router.currentRoute.value.fullPath === '/';
 });
-
-// Convert the object to an array of [key, value] pairs
-let languageOptionsArray = Object.entries(languageOptions);
-
-// Sort the array by the key (language code)
-languageOptionsArray.sort((a, b) => a[0].localeCompare(b[1]));
-
-// Convert it back to an object
-let sortedLanguageOptions = Object.fromEntries(languageOptionsArray);
-
-const languageDropdownOptions = computed(() => {
-  return Object.entries(sortedLanguageOptions).map(([key, value]) => {
-    return {
-      name: value.country,
-      code: value.code,
-      value: key,
-    };
-  });
-});
-
-const getCountryName = (locale) => {
-  return languageOptions[locale].country;
-};
-
-const getCountryFlag = (locale) => {
-  return languageOptions[locale].code;
-};
 
 const dropDownActions = computed(() => {
   const rawActions = getSidebarActions(isSuperAdmin.value, !isAtHome.value);
